+++ .husky/pre-commit
@@ -1,6 +1,3 @@
-#!/bin/sh
-. "$(dirname "$0")/_/husky.sh"
-
 echo "🔍 Running ESLint..."
 npm run lint || {
   echo "❌ Lint failed. Fix the issues before committing."
