+++ src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import App from './App.tsx';
-import ErrorBoundary from './components/ErrorBoundary.tsx';
+import App from './App/App.tsx';
+import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
 
 const root = document.getElementById('root');
 if (!root) throw new Error('Root element not found');
