+++ src/mocks/handlers.ts
@@ -1,5 +1,5 @@
 import { delay, http, HttpResponse } from 'msw';
-import type { SwapiPeopleResponse } from '../AppTypes';
+import type { SwapiPeopleResponse } from '../App/AppTypes';
 import { BASE_URL } from '../services/constants';
 
 export const mockCharacterResponse: SwapiPeopleResponse = {
